+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - for all measures, check whether the time step is valid in the function `validate_update_states_log`. If no, `NaN` is returned. Similar to `compute_criticality`: only compute for the valid ego vehicle and other vehicles, otherwise `NaN` is returned
 - check whether the vehicles are in the same lanelet: now the lanelet is extended by its successors and predecessors
 - adds the error handling of the out-of-projection-domain for many measures, such as TTC, THW, ALongReq, ALatReq
+- Initialization of the ego vehicle in the base class
 ## [0.3.2 & 0.3.3] - 2024.03.16
 ### Added
 - Error handling for check in same lanelet